update for split function from one
